+++ src/main/java/qubexplorer/ProjectNotFoundException.java
@@ -1,0 +1,18 @@
+package qubexplorer;
+
+/**
+ *
+ * @author Victor
+ */
+public class ProjectNotFoundException extends RuntimeException{
+    private final String shortProjectKey;
+
+    public ProjectNotFoundException(String shortProjectKey) {
+        this.shortProjectKey = shortProjectKey;
+    }
+
+    public String getShortProjectKey() {
+        return shortProjectKey;
+    }
+    
+}
