+++ src/main/java/qubexplorer/ui/CustomServerIssuesAction.java
@@ -11,7 +11,6 @@ import org.openide.windows.WindowManager;
 import qubexplorer.Summary;
 import qubexplorer.filter.IssueFilter;
 import qubexplorer.server.SonarQube;
-import qubexplorer.ui.options.SonarQubeOptionsPanel;
 import qubexplorer.ui.task.TaskExecutor;
 
 @ActionID(
