+++ src/main/resources/qubexplorer/ui/Bundle.properties
@@ -103,3 +103,10 @@ sonarqube-major-annotation=SonarQube Issue Major
 sonarqube-critical-annotation=SonarQube Issue Critical
 sonarqube-blocker-annotation=SonarQube Issue Blocker
 RuleDialog.ruleTitle.text_1=jLabel1
+SonarQubeOptionsPanel.jLabel1.text=Server Url:
+SonarQubeOptionsPanel.tJvmArguments.prompt=Example: -Xms500m -Xmx500m (options are separated by space)
+SonarQubeOptionsPanel.tJvmArguments.text=
+SonarQubeOptionsPanel.jLabel4.text=Runner JVM Arguments:
+SonarQubeOptionsPanel.jLabel3.text=(Analysis mode only applies in SonarQube 4.0 and above)
+SonarQubeOptionsPanel.jLabel2.text=Analysis Mode for Runner:
+SonarQubeOptionsPanel.serverAddress.text=
