--- src/main/resources/qubexplorer/ui/options/Bundle.properties
@@ -1,7 +1,0 @@
-SonarQubeOptionsPanel.jLabel1.text=Server Url:
-SonarQubeOptionsPanel.serverAddress.text=
-SonarQubeOptionsPanel.jLabel2.text=Analysis Mode for Runner:
-SonarQubeOptionsPanel.jLabel3.text=(Analysis mode only applies in SonarQube 4.0 and above)
-SonarQubeOptionsPanel.jLabel4.text=Runner JVM Arguments:
-SonarQubeOptionsPanel.tJvmArguments.prompt=Example: -Xms500m -Xmx500m (options are separated by space)
-SonarQubeOptionsPanel.tJvmArguments.text=
